requirements.txt
streamlit
fastapi
uvicorn
requests
sentence-transformers
chromadb
PyPDF2
transformers
huggingface_hub
langchain
google-generativeai
python-dotenv